Move clipboard to browser
